netstat: convert to sockaddr_storage
Rather than use different sockaddr base types for storage, always
start with sockaddr_storage on the stack and set up pointers to
more limited versions to that.  This way we always get the correct
alignment and storage when we pass it down to lower layers.

This is really just a set up for a follow up change to convert the
codebase entirely to sockaddr_storage.  This is just a fairly self
contained change.