ipmaddr: use safe_strncpy for ifr_name
This isn't strictly correct as the kernel doesn't require this
field be NUL terminated, but it matches how the rest of this
project handles ifr_name, so be consistent.  It also avoids a
new warning from gcc about unterminated buffers.

We should prob add a helper just for ifr_name to allow the
full width of the interface name when possible ...